Split ConnectionView, improve previews, attempt animation
